Bind buttons 2 and 3 appropriately on macOS aqua
Cf. https://rt.cpan.org/Ticket/Display.html?id=125050

macOS aqua uses buttons 2 and 3 for right and middle mouse buttons
respectively, which is the opposite of what X11 and Windows use.
As a workaround, check `windowingsystem` for 'aqua' to determine
which button to bind to.

Update references to "button 2" or "button 3" in documentation
to say "middle mouse button" and "right mouse button";
for consistency, change "button 1" to say "left mouse button".



git-svn-id: https://svn.code.sf.net/p/perl-tcl-ptk/code/trunk@213 8ab6fd2f-3ca1-42eb-bd32-3378fe2cd338